backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
boto3
reportlab
apscheduler
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
websockets
pandas
matplotlib
bcrypt==3.1.7
pydantic-settings
